[composables] Extract useAtMentionsMembers composable
Extract duplicated @-mention member list logic (reactive state,
filter function, team/taskTypes watchers) from AddComment and Comment
into a shared useAtMentionsMembers composable.
